+++ lib/rack/sw_webhook_authentication.rb
